Rename run_local -> callable_from_qua
